SystemUI: improve blur handling in respect to app launch animations
Once the animation starts, AOSP just disables blur. This works because
the QS panel is not transparent, so the user won't notice. Due to our
transparent QS panel patches, it becomes obvious that blur just
disappears once a user opens an app from QS (long press qs tile, use
settings button, etc). To improve this situation, we use a small trick:
Unlike how it looks in the animation, the shade is always on top of
apps, hence, our newly-launched app is below the shade and below the
blur. We disable blur in the very moment the app starts fading in to make
it as un-noticeable as possible for the user, without accidentally
blurring the app in the process.

Change-Id: Iecd794ef67c92a76f93c500e8feec9723f0e0d41
Signed-off-by: strongreasons <abenkenari@gmail.com>